+++ psRadiantOne/Public/DataSources/Get-R1DataSourceObject.ps1
@@ -1,0 +1,36 @@
+# .ExternalHelp psRadiantOne-help.xml
+function Get-R1DataSourceObject {
+	[CmdletBinding()]
+	[OutputType('System.String')]
+	param(
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[string]$name
+	)
+
+	Begin {
+
+		Assert-R1Session -RequireToken
+
+	}#begin
+
+	Process {
+
+		$URI = Resolve-R1ServiceUrl -Service Catalog -Path "data_sources/$($name | Get-EscapedString)/utils/objects"
+
+		$Result = Invoke-R1RestMethod -Uri $URI -Method GET
+
+		if ($null -ne $Result) {
+
+			$Result
+
+		}
+
+	}#process
+
+	End { }#end
+
+}